add aux files, copyright-notice

+++ reSID/filter.cc
@@ -228,7 +228,7 @@ void Filter::writeFC_LO(reg8 fc_lo)
 
 void Filter::writeFC_HI(reg8 fc_hi)
 {
-  fc = ((fc_hi << 3) & 0x7f8) | (fc & 0x007);
+  fc = (((unsigned int)fc_hi << 3) & 0x7f8) | (fc & 0x007);
   set_w0();
 }
